refactor: 인증 관련 패키지 분리 및 Basic Token 파싱 기능 병합
