+++ src/main/java/cart/exception/AuthLoginException.java
@@ -1,4 +1,4 @@
-package cart.config.auth;
+package cart.exception;
 
 public class AuthLoginException extends RuntimeException {
     public AuthLoginException(final String message) {
