+++ src/main/java/cart/exception/BasicAuthException.java
@@ -1,4 +1,4 @@
-package cart.config.auth;
+package cart.exception;
 
 public class BasicAuthException extends RuntimeException {
     public BasicAuthException(final String message) {
