+++ src/main/java/cart/service/request/AuthMember.java
@@ -1,4 +1,4 @@
-package cart.common.auth;
+package cart.service.request;
 
 import cart.domain.member.MemberId;
